pcoc: Run spatch in parallel
pcoc (pronounced 'peacock') is a simple script for running a single
coccinelle script in parallel, using GNU parallel to distribute the
workload over several spatch instances.

The linux kernel already contains a similar facility in the form of
the coccicheck script, invoked via the same-name Make
target. Unfortunately, I've seen garbled output a number of times,
since all the spatch instances write to the same stdout (whether
that's connected to the terminal, a file or a pipe). Also, it is a
little annoying that the output from various subsystems get mixed up -
I often save the output to a file and then do 'git apply --stat' to
see where it makes most sense to actually apply the patches, but this
is hard when the output then looks like

 mm/page_alloc.c                                    |    3 +--
 drivers/net/wireless/ath/ath9k/mci.c               |   12 ++++--------
 drivers/mmc/host/msm_sdcc.c                        |    3 +--
 drivers/net/wireless/ath/regd.c                    |    3 +--
 fs/xfs/xfs_trans_buf.c                             |    4 +---
 drivers/net/wireless/ath/ath5k/base.c              |    3 +--
 ...

These are problems which GNU parallel has already solved pretty well,
so I decided to write a script based on that.

Also, I almost always just want to run a single coccinelle script,
possibly restricted to some subsystem. So I decided that the syntax
should be as simple as possible for that case. The coccicheck
invocation

  make coccicheck COCCI=file.cocci M=fs/ MODE=patch

can mostly be replaced by

  pcoc --mode=patch file.cocci fs

and the --mode=patch part is not even necessary if patch is the
'primary' mode in file.cocci (that is, patch is the first virtual rule
declared).

It still needs some polishing, such as extracting the suggested commit
message when in patch mode, but it should already be usable (also
outside the linux kernel, btw).