round_up.cocci
Rounding x to the next even integer can be done in a variety of
ways. Usually the simplest and most readable is

  x = ALIGN(x, 2)

since that expands to x = (x+1)&~1, removing the need for extra registers.

Another way is

  x += x & 1;

which gcc currently fails to optimize to the above, so it requires an
extra register and an extra mov instruction. Then we also have

  if (x & 1)
    ++x;

which is particularly bad.

While at it try to check for aligning to higher powers of 2. I've seen
a few instances of "x += x&1; x += x&2;", and they may creep back in.