test_and_setbit.cocci: Improve suggested commit message
Looking at some of the instances found by the coccinelle script, I
found an example of what I suspected might exist: A valid reason for
testing the bit before setting it. So mention this fact. But also
emphasize that such cases are the exception rather than the rule.

Also, spatch seems to perform rather poorly when using lots of
alternatives, so simply spell out all the variations I could come up
with. Use X == 0 instead of !X since there's a standard isomorphism
making the former also match the latter. For some reason, this also
seems to improve the matching speed.