snprintf.cocci: Replace snprintf with strlcpy
I'm not a huge fan of strlcpy, but it's quite a bit cheaper than
snprintf for copying a literal string. They have exactly the same
semantics (ignoring the different return types): Guarantee
nul-termination of the destination (unless given size==0) and return
the length of the source string, not counting the terminating '\0'.