strncmp_*: Check some strncmp-related issues
Checking whether a string starts with "abc" is often done using the
idiom !strncmp(s, "abc", 3). However, in a few cases the hard-coded
length is wrong. It would be better to spell it !strncmp(s, "abc",
strlen("abc")). But this is still suboptimal, since the literal is
needlessly duplicated. strstarts() exists for this purpose. The
coccinelle script strncmp_strstarts produces patches changing three
common patterns into strstarts:

  !strncmp(s, "abc", strlen("abc"))
  !strncmp(s, "abc", sizeof("abc") - 1)
  !strncmp(s, "abc", 3)

The last case is fed through a little embedded python to check that
the length is indeed the length of the literal.

To deal with the cases where the length is different, there are two
other coccinelle scripts. strncmp_strcmp.cocci suggests replacing
strncmp(s, "abc", sizeof("abc")) with strcmp(s, "abc") [note the lack
of ! on these; this works more generally than the strstarts
replacement], in particular because the former is easily confused with
the above 'is "abc" a prefix of s' case.

Finally, there are cases where the length is silly, such as strncmp(s,
"abc", 1). In that case, one might as well pass the literal "a", but
it is more likely that the length is simply wrong. strncmp_bad.cocci
points out such cases (but obviously cannot suggest a patch; it is not
clear whether one wanted to test for equality or prefix or something
else).

All of the above also applies to the case-insensitive case. Although
strcasestarts() doesn't exist, it should be uncontroversial to
introduce it.