self_assign.cocci: Some more WTFs
Assigning a variable to itself is rarely useful. Unfortunately, it is
commonly used for local variables to silence a 'may be used
uninitialized' warning, so we don't chech for that case. But other
instances where the lvalue is part of some struct, or if the operator
is &= or |=, often indicate a copy-paste error.

We also check for some common str/mem functions being called with two
identical operands, which is almost certainly a bug.