redundant_casecmp: Replace !strcasecmp with !strcmp
If a string literal doesn't contain any letters, a string is equal to
the literal modulo case if and only if it is actually equal to it.

For example, '!strcasecmp(f, "1")' can be replaced by '!strcmp(f,
"1")'. Aside from being slighlty cheaper (strcasecmp must spend cycles
converting to lowercase, and most likely incurs a cache miss on the
first call to tolower()), gcc can completely eliminate string
comparisons to short strings. It will compile this example to
something like 'f[0] == '1' && !f[1]', since it knows the semantics of
strcmp (including that f[0] is always safe to access, and if that is
non-zero, f[1] is also safe to access).