Find vararg functions
This was just a quick hack used to extract a rough list of functions
taking varargs. I needed that for estimating the effect of adding a
flag to gcc to suppress emitting 'xor %eax, %eax' before each and
every call of a variadic function. It turned out a defconfig kernel
contained 13000+ such calls, meaning that the potential saving was
over 25kB of .text. After I actually suggested it, it took H.J. Lu
less than a day to implement it and, modulo some documentation issues,
getting it into gcc trunk.