copy_bit.cocci: a micro-optimization
if (a & FLAG)
  b |= FLAG;

can usually be transformed to the branch-free

b |= a & FLAG;

However, gcc doesn't do this, because it is illegal if b is not
writable unless FLAG is set in a. But that would be a rather unusual
situation, and certainly isn't an issue in the common case where b is
a stack variable or has just been written to unconditionally.