Replace strncpy with strlcpy
Doing

  strncpy(dst, src, sizeof(dst));
  dst[sizeof(dst)-1] = '\0';

can usually be replaced by

  strlcpy(dst, src, sizeof(dst));

They are not exactly equivalent, but one is very rarely interested
in the 'fill the rest of the buffer, if any, with 0' "feature" of
strncpy. In fact, that's usually just an annoying waste of cycles.

Feel free to replace strlcpy by strscpy if you think that's nicer.