wtf/bad_delay.cocci: find no-op "delay loops"
gcc -O2 optimizes a do-nothing loop like

for (i=0; i<10000; ++i);

into

i=10000;

One needs to be a little more fancy. One way is to declare i volatile,
another is to just do it in inline asm. A third is to just delete the
code, since if it's been working till now, no delay is apparently
needed.